feat: change file structure cli web-components template
