+++ code/renderers/web-components/template/cli/js/Page/index.js
@@ -1,5 +1,5 @@
 import { html } from 'lit';
-import { Header } from './Header';
+import { Header } from '../Header';
 import './page.css';
 
 export const Page = ({ user, onLogin, onLogout, onCreateAccount }) => html`
