+++ code/renderers/web-components/template/cli/ts-4-9/Header/Header.stories.ts
@@ -1,13 +1,13 @@
 import type { Meta, StoryObj } from '@storybook/web-components';
-import type { HeaderProps } from './Header';
-import { Header } from './Header';
+import type { HeaderProps } from '.';
+import { Header } from '.';
 
-const meta: Meta<HeaderProps> = {
+const meta = {
   title: 'Example/Header',
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/web-components/writing-docs/autodocs
   tags: ['autodocs'],
   render: (args: HeaderProps) => Header(args),
-};
+} satisfies Meta<HeaderProps>;
 
 export default meta;
 type Story = StoryObj<HeaderProps>;
@@ -15,7 +15,7 @@ type Story = StoryObj<HeaderProps>;
 export const LoggedIn: Story = {
   args: {
     user: {
-      name: 'Jonh Doe',
+      name: 'Jane Doe',
     },
   },
 };
